Add files via upload
A hypergraph H is a pair (V, E) where V is a set of elements called vertices, and E is a set of non-empty subsets of V called edges. Therefore, E is a subset of P(V ) {φ}, where P(V ) is the power set of V . (Directed Hypergraph) 
A hypergraph H = (V, E) is called a directed hypergraph if it satisfies the following properties: 1. each edge is a directed edge and 2. for any two edges e1, e2 ∈ E, T e1 ∪ He16 = T e2 ∪ He2 A directed hypergraph H = (V, E) is called K-uniform, ∥e∥ = K, ∀e ∈ E. A directed hypergraph that is not uniform is called a directed non-uniform hypergraph.
The main idea behind our HyperGen is to take the uniform and non-uniform interactions in hypergraphs into consideration. Intuitively, in order to reproduce the desired patterns in multilevel decomposed graphs, the generation process should have the following characteristics: Get Input Parameters: While generating a hypergraph, the HyperGen takes the following input parameters: i) the number of nodes, ii) the number of hypergraphs, iii) the number of hyperedges, iv the number of classes, v) the number of nodes per hyperedge. The number of vertices (parameter i), number of hypergraphs (parameter ii), and number of hyperedges (parameter iii) are the maximum number of vertices and hyperedges a generated hypergraph can contain.